Separate function-domain folder
